Fixed circular dependency, but feels somewhat ugly
